feat: Refactor status handling by consolidating getStatusColor and getStatusText into utils, and remove unused code from components
